+++ example/resource_virtual_environment_time.tf
@@ -1,0 +1,12 @@
+resource "proxmox_virtual_environment_time" "example" {
+  node_name = "${data.proxmox_virtual_environment_time.example.node_name}"
+  time_zone = "${data.proxmox_virtual_environment_time.example.time_zone}"
+}
+
+output "resource_proxmox_virtual_environment_time" {
+  value = "${map(
+    "local_time", proxmox_virtual_environment_time.example.local_time,
+    "time_zone", proxmox_virtual_environment_time.example.time_zone,
+    "utc_time", proxmox_virtual_environment_time.example.utc_time,
+  )}"
+}